Move tests for nomis and ndelius IDs into service test file.
