+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppssubjectaccessrequestapi/controllers/SubjectAccessRequestControllerTest.kt
@@ -32,105 +32,19 @@ class SubjectAccessRequestControllerTest {
     "ndeliusId: '1' " +
     "}"
 
-  private val ndeliusAndNomisRequest = "{ " +
-    "dateFrom: '01/12/2023', " +
-    "dateTo: '03/01/2024', " +
-    "sarCaseReferenceNumber: '1234abc', " +
-    "services: '{1,2,4}', " +
-    "nomisId: '1', " +
-    "ndeliusId: '1' " +
-    "}"
-
-  private val noIDRequest = "{ " +
-    "dateFrom: '01/12/2023', " +
-    "dateTo: '03/01/2024', " +
-    "sarCaseReferenceNumber: '1234abc', " +
-    "services: '{1,2,4}', " +
-    "nomisId: '', " +
-    "ndeliusId: '' " +
-    "}"
-
-  private val json = JSONObject(ndeliusRequest)
-  private val formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy")
-  private val dateFrom = json.get("dateFrom").toString()
-  private val dateFromFormatted = LocalDate.parse(dateFrom, formatter)
-  private val dateTo = json.get("dateTo").toString()
-  private val dateToFormatted = LocalDate.parse(dateTo, formatter)
   private val requestTime = LocalDateTime.now()
-  private val sampleSAR = SubjectAccessRequest(
-    id = null,
-    status = Status.Pending,
-    dateFrom = dateFromFormatted,
-    dateTo = dateToFormatted,
-    sarCaseReferenceNumber = "1234abc",
-    services = "{1,2,4}",
-    nomisId = "",
-    ndeliusCaseReferenceId = "1",
-    requestedBy = "aName",
-    requestDateTime = requestTime,
-    claimAttempts = 0,
-  )
-
-  private val sampleUnclaimedSAR = SubjectAccessRequest(
-    id = null,
-    status = Status.Pending,
-    dateFrom = dateFromFormatted,
-    dateTo = dateToFormatted,
-    sarCaseReferenceNumber = "1234abc",
-    services = "{1,2,4}",
-    nomisId = "",
-    ndeliusCaseReferenceId = "1",
-    requestedBy = "Test",
-    requestDateTime = requestTime,
-    claimAttempts = 0,
-  )
+  private val sarService = Mockito.mock(SubjectAccessRequestService::class.java)
+  private val auditService = Mockito.mock(AuditService::class.java)
+  private val authentication: Authentication = Mockito.mock(Authentication::class.java)
   @Test
-  fun `createSubjectAccessRequestPost calls service createSubjectAccessRequestPost and returns 200`() {
-    val sarRepository = Mockito.mock(SubjectAccessRequestRepository::class.java)
-    val sarService = Mockito.mock(SubjectAccessRequestService::class.java)
-    val auditService = Mockito.mock(AuditService::class.java)
-    val authentication: Authentication = Mockito.mock(Authentication::class.java)
+  fun `createSubjectAccessRequestPost calls service createSubjectAccessRequestPost with same parameters`() {
     Mockito.`when`(authentication.name).thenReturn("aName")
-    val expected = ResponseEntity("", HttpStatus.OK)
-    val result: ResponseEntity<String> = SubjectAccessRequestController(sarService, auditService)
+    SubjectAccessRequestController(sarService, auditService)
       .createSubjectAccessRequestPost(ndeliusRequest, authentication, requestTime)
     verify(sarService, times(1)).createSubjectAccessRequestPost(ndeliusRequest, authentication, requestTime)
-    Assertions.assertThat(result).isEqualTo(expected)
   }
-
-  @Test
-  fun `createSubjectAccessRequestPost returns 400 and error string if both IDs are supplied`() {
-    val sarRepository = Mockito.mock(SubjectAccessRequestRepository::class.java)
-    val auditService = Mockito.mock(AuditService::class.java)
-    val sarService = Mockito.mock(SubjectAccessRequestService::class.java)
-    val authentication: Authentication = Mockito.mock(Authentication::class.java)
-    Mockito.`when`(authentication.name).thenReturn("aName")
-    val expected = ResponseEntity("Both nomisId and ndeliusId are provided - exactly one is required", HttpStatus.BAD_REQUEST)
-    val result: ResponseEntity<String> = SubjectAccessRequestController(sarService, auditService)
-      .createSubjectAccessRequestPost(ndeliusAndNomisRequest, authentication, requestTime)
-    verify(sarService, times(1)).createSubjectAccessRequestPost(ndeliusAndNomisRequest, authentication, requestTime)
-    Assertions.assertThat(result).isEqualTo(expected)
-  }
-
-  @Test
-  fun `createSubjectAccessRequestPost returns 400 and error string if neither ID is supplied`() {
-    val sarRepository = Mockito.mock(SubjectAccessRequestRepository::class.java)
-    val auditService = Mockito.mock(AuditService::class.java)
-    val sarService = Mockito.mock(SubjectAccessRequestService::class.java)
-    val authentication: Authentication = Mockito.mock(Authentication::class.java)
-    Mockito.`when`(authentication.name).thenReturn("aName")
-    val expected = ResponseEntity("Neither nomisId nor ndeliusId is provided - exactly one is required", HttpStatus.BAD_REQUEST)
-    val result: ResponseEntity<String> = SubjectAccessRequestController(sarService, auditService)
-      .createSubjectAccessRequestPost(noIDRequest, authentication, requestTime)
-    verify(sarService, times(1)).createSubjectAccessRequestPost(noIDRequest, authentication, requestTime)
-    Assertions.assertThat(result).isEqualTo(expected)
-  }
-
   @Test
   fun `getSubjectAccessRequests is called with unclaimedOnly = true if specified in controller and returns list`() {
-    val sarRepository = Mockito.mock(SubjectAccessRequestRepository::class.java)
-    val auditService = Mockito.mock(AuditService::class.java)
-    val sarService = Mockito.mock(SubjectAccessRequestService::class.java)
     val result: List<SubjectAccessRequest?> = SubjectAccessRequestController(sarService, auditService)
       .getSubjectAccessRequests(unclaimed = true)
     verify(sarService, times(1)).getSubjectAccessRequests(unclaimedOnly = true)
@@ -139,10 +53,7 @@ class SubjectAccessRequestControllerTest {
 
   @Test
   fun `getSubjectAccessRequests is called with unclaimedOnly = false if unspecified in controller`() {
-    val sarRepository = Mockito.mock(SubjectAccessRequestRepository::class.java)
-    val auditService = Mockito.mock(AuditService::class.java)
-    val sarService = Mockito.mock(SubjectAccessRequestService::class.java)
-    val result: List<SubjectAccessRequest?> = SubjectAccessRequestController(sarService, auditService)
+    SubjectAccessRequestController(sarService, auditService)
       .getSubjectAccessRequests()
     verify(sarService, times(1)).getSubjectAccessRequests(unclaimedOnly = false)
   }
